+++ cmdstanpy/__init__.py
@@ -1,15 +1,7 @@
 import atexit
-import os
-import os.path
 import shutil
 import tempfile
 
-if 'CMDSTAN' in os.environ:
-    CMDSTAN_PATH = os.environ['CMDSTAN']
-else:
-    CMDSTAN_PATH = os.path.abspath(os.path.join('.', 'releases', 'cmdstan'))
-
-TMPDIR = tempfile.mkdtemp()
 STANSUMMARY_STATS = [
     'Mean',
     'MCSE',
@@ -21,6 +13,8 @@ STANSUMMARY_STATS = [
     'N_Eff/s',
     'R_hat',
 ]
+
+TMPDIR = tempfile.mkdtemp()
 
 
 def cleanup_tmpdir():
